[Stdlib] Add two new intrinsic functions for assertion
Two new intrinsic functions were added
1. `isAssertionArgumentEvaluationEnabled`
Allows evaluating conditions in `assert` call.
2. `isAssertionThrowingExceptionEnabled`
Allows throwing `AssertionError` if the condition of `assert` was false.

These new intrinsics are required because of the new inlining model.
`assert` is an inline function. In the new model the inlined version
of a function will be written in klib. There will be no way to determine
if this particular code came from the `assert` call. But we still
want to be able to enable/disable assertions. So we are going to wrap
the assertion call inside an `if` statement with these new intrinsics
as conditions. On the second stage of compilation (after all klibs
are loaded), these intrinsics will be replaced with actual values
depending on the assertion mode.

^KT-69167 Fixed
^KT-67292