[IR] Don't try to interpret an already inlined function
Earlier IR interpreter tried to evaluate functions that were inlined
based on their original inline call. For example, `Char.code`.
In Native backend, an inlined version of this extension property
has some Native-specific calls that IR interpreter can't evaluate. So
the original call was used instead, but this original call will
not be available later on after we move inliner on the first phase.
This means that the compiler can't optimize `Char.code`
expression (in bodies), but this shouldn't be a big deal.

^KT-69467