[FIR] Forbid using an intersection type for reified type parameter
When an intersection type is inferred and used for a reified type
parameter, subtle bugs may occur becase the reified type resolves to the
upper bound of the intersection (often Any?). This may cause unexpected
behavior if the type is used for instantiating a class like in Kotlin
serialization.

Instead of having un predictable behavior at runtime, issue a warning on
the type parameter, so the user may explicitly specify the type desired.

^KT-52469 Fixed