[FIR] Downgrade `RedundantExplicitTypeChecker` to experimental checkers
The checker produces a false positive on integer literals:

```
val it: Long = 0
```

Looks like we don't have the information about how the literal looked
like initially: we have `kind`, but it's replaced when we fix the
literal variable to a specific type with the actual type information,
thus duplicating it. The fix probably deserves a separate refactoring.