JVM: fix IR -> PSI mapping for delegated property accessors in K2
In K2, the delegate expression is used as the source of the property,
instead of the encompassing "by ..." expression. So we support both K1
and K2 cases in JvmIrDeclarationOrigin.

Another way to fix this would be to replicate K1 behavior in
`PsiRawFirBuilder.Visitor.toFirProperty`, however that change affects
far more places in diagnostic reporting and it isn't obvious that it is
better, while this fix is local.

This doesn't affect any existing tests, but fixes for example
testDelegateCorrectErrorTypes in the upcoming K2 kapt implementation
(KT-69860) because K2 kapt is now able to find the PSI element and
correct the error type.