Resolve Fir2IrLazyClass for annotation call on convertToIrConstructorCall
The compose compiler plugin accesses the `Fir2IrLazyClass::declarations`
in an IR lowering transformer. When the class for `Fir2IrLazyClass`
calls an annotation constructor with arguments, unresolved
`FirAnnotationCall` results in a wrong value argument mapping. This
commit fixes it by running `lazyResolveToPhase(..)` for the containing
symbol for the annoation constructor call on
`convertToIrConstructorCall(..)`.

^KT-70545